Refactored dropCount to use int instead of double
This change ensures consistency in data types by converting dropCount from double to int. It prevents potential type-related issues and simplifies handling of this value in resupply logic.

+++ MekHQ/src/mekhq/campaign/mission/resupplyAndCaches/PerformResupply.java
@@ -98,7 +98,7 @@ public class PerformResupply {
      * @param dropCount the number of supply drops planned for this resupply operation. If zero,
      *                  the method exits early.
      */
-    public static void performResupply(Resupply resupply, AtBContract contract, double dropCount) {
+    public static void performResupply(Resupply resupply, AtBContract contract, int dropCount) {
         // These early exits should only occur if the player literally has no units.
         if (dropCount == 0) {
             logger.info("Resupply exited early, as DropCount is 0");
